Refactor order color settings example and cleanup props
Moved the order color settings example table to a new DisplayExampleTable component and removed related dummy data and props from DefaultOrdersTable and OrderColorSettings. Cleaned up unused or redundant props (such as isSetting and isOrderColorSetting) across related components, and improved type imports for consistency. This refactor improves code clarity and separation of concerns for ICU order color settings.
