+++ components/hospital/common/default-template-order/dt-table-header.tsx
@@ -2,9 +2,10 @@ import OrderWidthButton from '@/components/hospital/icu/main/chart/selected-char
 import SortingButton from '@/components/hospital/icu/main/chart/selected-chart/chart-body/table/chart-table-header/sorting-button'
 import { TableHead, TableHeader, TableRow } from '@/components/ui/table'
 import { TIMES } from '@/constants/hospital/icu/chart/time'
-import { type OrderWidth } from '@/types/hospital/order'
-import { type SelectedIcuOrder } from '@/types/icu/chart'
+import type { OrderWidth } from '@/types/hospital/order'
+import type { SelectedIcuOrder } from '@/types/icu/chart'
 import type { Dispatch, SetStateAction } from 'react'
+
 type Props = {
   isSorting: boolean
   setIsSorting: Dispatch<SetStateAction<boolean>>
@@ -12,7 +13,6 @@ type Props = {
   setOrderWidth: Dispatch<SetStateAction<OrderWidth>>
   sortedOrders: SelectedIcuOrder[]
   defaultChartOrders: SelectedIcuOrder[]
-  isSetting?: boolean
 }
 
 export default function DtTableHeader({
@@ -22,7 +22,6 @@ export default function DtTableHeader({
   setIsSorting,
   sortedOrders,
   defaultChartOrders,
-  isSetting,
 }: Props) {
   return (
     <TableHeader className="shadow-sm">
@@ -40,7 +39,6 @@ export default function DtTableHeader({
             isSorting={isSorting}
             setIsSorting={setIsSorting}
             isDt
-            isSetting={isSetting}
           />
 
           <span className="text-center">기본오더</span>
